move benchmarking stuff to benchmarking module
