+++ system-parachains/encointer/src/impls.rs
@@ -22,9 +22,12 @@ pub mod benchmarks {
 	use core::marker::PhantomData;
 	use cumulus_primitives_core::{ChannelStatus, GetChannelInfo};
 	use frame_support::traits::tokens::PaymentStatus;
-	use pallet_encointer_treasuries::Transfer;
-	use sp_core::Get;
-	use xcm::{latest::Location, opaque::latest::Junction::Parachain, GetVersion};
+	use pallet_encointer_treasuries::{
+		benchmarking::ArgumentsFactory as TreasuryArgumentsFactory, Transfer,
+	};
+	use polkadot_runtime_common::impls::VersionedLocatableAsset;
+	use sp_core::{ConstU32, ConstU8, Get};
+	use xcm::{latest::Location, opaque::latest::Junction::Parachain, prelude::*, GetVersion};
 
 	/// Trait for setting up any prerequisites for successful execution of benchmarks.
 	pub trait EnsureSuccessful {
@@ -98,6 +101,30 @@ pub mod benchmarks {
 		}
 		fn ensure_concluded(id: Self::Id) {
 			O::ensure_concluded(id)
+		}
+	}
+
+	// The below implementation is basically the same as for:
+	// polkadot_runtime_common::impls::benchmarks::TreasuryArguments
+
+	/// Provide factory methods for the [`VersionedLocatableAsset`].
+	/// The location of the asset is determined as a Parachain with an
+	/// ID equal to the passed seed.
+	pub struct TreasuryArguments<Parents = ConstU8<0>, ParaId = ConstU32<0>>(
+		PhantomData<(Parents, ParaId)>,
+	);
+	impl<Parents: Get<u8>, ParaId: Get<u32>> TreasuryArgumentsFactory<VersionedLocatableAsset>
+		for TreasuryArguments<Parents, ParaId>
+	{
+		fn create_asset_kind(seed: u32) -> VersionedLocatableAsset {
+			(
+				Location::new(Parents::get(), [Junction::Parachain(ParaId::get())]),
+				AssetId(Location::new(
+					0,
+					[PalletInstance(seed.try_into().unwrap()), GeneralIndex(seed.into())],
+				)),
+			)
+				.into()
 		}
 	}
 }
