+++ system-parachains/encointer/src/lib.rs
@@ -629,41 +629,7 @@ impl pallet_encointer_treasuries::Config for Runtime {
 	>;
 	#[cfg(feature = "runtime-benchmarks")]
 	type BenchmarkHelper =
-		benchmarks_helper::TreasuryArguments<sp_core::ConstU8<1>, ConstU32<1000>>;
-}
-
-#[cfg(feature = "runtime-benchmarks")]
-pub mod benchmarks_helper {
-	use super::VersionedLocatableAsset;
-	use core::marker::PhantomData;
-	use frame_support::traits::Get;
-	use pallet_encointer_treasuries::benchmarking::ArgumentsFactory as TreasuryArgumentsFactory;
-	use sp_core::{ConstU32, ConstU8};
-	use xcm::prelude::*;
-
-	// The below implementation is basically the same as for:
-	// polkadot_runtime_common::impls::benchmarks::TreasuryArguments
-
-	/// Provide factory methods for the [`VersionedLocatableAsset`].
-	/// The location of the asset is determined as a Parachain with an
-	/// ID equal to the passed seed.
-	pub struct TreasuryArguments<Parents = ConstU8<0>, ParaId = ConstU32<0>>(
-		PhantomData<(Parents, ParaId)>,
-	);
-	impl<Parents: Get<u8>, ParaId: Get<u32>> TreasuryArgumentsFactory<VersionedLocatableAsset>
-		for TreasuryArguments<Parents, ParaId>
-	{
-		fn create_asset_kind(seed: u32) -> VersionedLocatableAsset {
-			(
-				Location::new(Parents::get(), [Junction::Parachain(ParaId::get())]),
-				AssetId(Location::new(
-					0,
-					[PalletInstance(seed.try_into().unwrap()), GeneralIndex(seed.into())],
-				)),
-			)
-				.into()
-		}
-	}
+		impls::benchmarks::TreasuryArguments<sp_core::ConstU8<1>, ConstU32<1000>>;
 }
 
 impl pallet_aura::Config for Runtime {
